backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
neo4j==5.25.0
confluent-kafka
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
